refactor!: rename UniDirectional namespace to OneWay
