+++ Testing/CarbonatePerfTests/PtrDataPuller.cs
@@ -5,7 +5,7 @@
 namespace CarbonatePerfTests;
 
 using System.Runtime.CompilerServices;
-using Carbonate.UniDirectional;
+using Carbonate.OneWay;
 
 /// <summary>
 /// Used for perf testing.
