+++ Testing/CarbonateTests/OneWay/PullReactableTests.cs
@@ -2,10 +2,10 @@
 // Copyright (c) KinsonDigital. All rights reserved.
 // </copyright>
 
-namespace CarbonateTests.UniDirectional;
+namespace CarbonateTests.OneWay;
 
-using Carbonate.Core.UniDirectional;
-using Carbonate.UniDirectional;
+using Carbonate.Core.OneWay;
+using Carbonate.OneWay;
 using FluentAssertions;
 using Moq;
 using Xunit;
