+++ Testing/CarbonateTests/OneWay/ReceiveReactorTests.cs
@@ -2,9 +2,9 @@
 // Copyright (c) KinsonDigital. All rights reserved.
 // </copyright>
 
-namespace CarbonateTests.UniDirectional;
+namespace CarbonateTests.OneWay;
 
-using Carbonate.UniDirectional;
+using Carbonate.OneWay;
 using FluentAssertions;
 using Xunit;
